+++ src/Listeners/SwitchPreferredTranslations.php
@@ -2,14 +2,14 @@
 
 namespace Nevadskiy\Translatable\Listeners;
 
-use Nevadskiy\Translatable\Events\TranslationSavedEvent;
+use Nevadskiy\Translatable\Events\TranslationSaved;
 
 class SwitchPreferredTranslations
 {
     /**
      * Handle the given event.
      */
-    public function handle(TranslationSavedEvent $event): void
+    public function handle(TranslationSaved $event): void
     {
         $event->translation->translatable->translations()
             ->whereKeyNot($event->translation->id)