+++ src/TranslatableServiceProvider.php
@@ -5,7 +5,7 @@ namespace Nevadskiy\Translatable;
 use Illuminate\Contracts\Events\Dispatcher;
 use Illuminate\Foundation\Events\LocaleUpdated;
 use Illuminate\Support\ServiceProvider;
-use Nevadskiy\Translatable\Events\TranslationSavedEvent;
+use Nevadskiy\Translatable\Events\TranslationSaved;
 
 class TranslatableServiceProvider extends ServiceProvider
 {
@@ -19,7 +19,7 @@ class TranslatableServiceProvider extends ServiceProvider
             Listeners\UpdateTranslatorLocale::class,
         ],
 
-        TranslationSavedEvent::class => [
+        TranslationSaved::class => [
             Listeners\SwitchPreferredTranslations::class,
         ],
     ];