+++ tests/Feature/FireTranslationSavedTest.php
@@ -3,7 +3,7 @@
 namespace Nevadskiy\Translatable\Tests\Feature;
 
 use Illuminate\Support\Facades\Event;
-use Nevadskiy\Translatable\Events\TranslationSavedEvent;
+use Nevadskiy\Translatable\Events\TranslationSaved;
 use Nevadskiy\Translatable\Tests\Support\Factories\BookFactory;
 use Nevadskiy\Translatable\Tests\TestCase;
 
@@ -14,11 +14,11 @@ class FireTranslationSavedTest extends TestCase
     {
         $book = BookFactory::new()->create();
 
-        Event::fake(TranslationSavedEvent::class);
+        Event::fake(TranslationSaved::class);
 
         $book->translate('title', 'Моя книга', 'ru');
 
-        Event::assertDispatched(TranslationSavedEvent::class, static function (TranslationSavedEvent $event) use ($book) {
+        Event::assertDispatched(TranslationSaved::class, static function (TranslationSaved $event) use ($book) {
             return $event->translation->translatable->is($book)
                 && $event->translation->translatable_attribute === 'title'
                 && $event->translation->locale === 'ru'
@@ -29,10 +29,10 @@ class FireTranslationSavedTest extends TestCase
     /** @test */
     public function it_does_not_fire_translation_saved_event_when_translatable_model_is_just_created(): void
     {
-        Event::fake(TranslationSavedEvent::class);
+        Event::fake(TranslationSaved::class);
 
         BookFactory::new()->create();
 
-        Event::assertNotDispatched(TranslationSavedEvent::class);
+        Event::assertNotDispatched(TranslationSaved::class);
     }
 }